some immprovemts,  more coming soon...
