+++ .fleet/run.json
@@ -1,0 +1,11 @@
+{
+    "configurations": [
+        {
+            "type": "python",
+            "name": "Python configuration",
+            "file": "part_1_the_beginning.py",
+            "arguments": [],
+        },
+        
+    ]
+}
